#1702: Move to 'startup' package.
