--- cmd/common/Attribution.txt
@@ -1,142 +1,0 @@
-The following open source projects are referenced by Core Common:
-
-pkg/errors (BSD-2) https://github.com/pkg/errors
-https://github.com/pkg/errors/blob/master/LICENSE
-
-gorilla/mux (BSD-3) https://github.com/gorilla/mux
-https://github.com/gorilla/mux/blob/master/LICENSE
-
-globalsign/mgo (unspecified) - https://github.com/globalsign/mgo
-https://github.com/globalsign/mgo/blob/master/LICENSE
-
-pebbe/zmq4 (BSD-2) https://github.com/pebbe/zmq4
-https://github.com/pebbe/zmq4/blob/master/LICENSE.txt
-
-go-kit/kit (MIT) github.com/go-kit/kit
-https://github.com/go-kit/kit/blob/master/LICENSE
-
-go-logfmt/logfmt (MIT) https://github.com/go-logfmt/logfmt
-https://github.com/go-logfmt/logfmt/blob/master/LICENSE
-
-robfig/cron (MIT) https://github.com/robfig/cron
-https://github.com/robfig/cron/blob/master/LICENSE
-
-dgrijalva/jwt-go (MIT) https://github.com/dgrijalva/jwt-go
-https://github.com/dgrijalva/jwt-go/blob/master/LICENSE
-
-google/uuid (BSD-3) https://github.com/google/uuid
-https://github.com/google/uuid/blob/master/LICENSE
-
-pelletier/go-toml (MIT) https://github.com/pelletier/go-toml
-https://github.com/pelletier/go-toml/blob/master/LICENSE
-
-influxdata/influxdb/client/v2 (MIT) https://github.com/influxdata/influxdb
-https://github.com/influxdata/influxdb/blob/master/LICENSE
-
-influxdata/platform (MIT) https://github.com/influxdata/platform
-https://github.com/influxdata/platform/blob/master/LICENSE
-
-eclipse/paho.mqtt.golang (Eclipse Public License 1.0) https://github.com/eclipse/paho.mqtt.golang
-https://github.com/eclipse/paho.mqtt.golang/blob/master/LICENSE
-
-mattn/go-xmpp (BSD-3) https://github.com/mattn/go-xmpp
-https://github.com/mattn/go-xmpp/blob/master/LICENSE
-
-BurntSushi/toml (MIT) https://github.com/BurntSushi/toml
-https://github.com/BurntSushi/toml/blob/master/COPYING
-
-mitchellh/consulstructure (MIT) https://github.com/mitchellh/consulstructure
-https://github.com/mitchellh/consulstructure/blob/master/LICENSE
-
-mitchellh/mapstructure (MIT) https://github.com/mitchellh/mapstructure
-https://github.com/mitchellh/mapstructure/blob/master/LICENSE
-
-mitchellh/copystructure (MIT) https://github.com/mitchellh/copystructure
-https://github.com/mitchellh/copystructure/blob/master/LICENSE
-
-mitchellh/reflectwalk (MIT) https://github.com/mitchellh/reflectwalk
-https://github.com/mitchellh/reflectwalk/blob/master/LICENSE
-
-cenkalti/backoff (MIT) https://github.com/cenkalti/backoff
-https://github.com/cenkalti/backoff/blob/master/LICENSE
-
-hashicorp/consul 1.1.0 (Mozilla Public License 2.0) - https://github.com/hashicorp/consul
-https://github.com/hashicorp/consul/blob/master/LICENSE
-
-hashicorp/go-cleanhttp (Mozilla Public License 2.0) - https://github.com/hashicorp/go-cleanhttp
-https://github.com/hashicorp/go-cleanhttp/blob/master/LICENSE
-
-hashicorp/go-rootcerts (Mozilla Public License 2.0) https://github.com/hashicorp/go-rootcerts
-https://github.com/hashicorp/go-rootcerts/blob/master/LICENSE
-
-mitchellh/go-homedir (MIT) https://github.com/mitchellh/go-homedir
-https://github.com/mitchellh/go-homedir/blob/master/LICENSE
-
-mitchellh/mapstructure (MIT) https://github.com/mitchellh/mapstructure
-https://github.com/mitchellh/mapstructure/blob/master/LICENSE
-
-hashicorp/serf (Mozilla Public License 2.0) https://github.com/hashicorp/serf
-https://github.com/hashicorp/serf/blob/master/LICENSE
-
-armon/go-metrics (MIT) https://github.com/armon/go-metrics
-https://github.com/armon/go-metrics/blob/master/LICENSE
-
-hashicorp/go-immutable-radix (Mozilla Public License 2.0) https://github.com/hashicorp/go-immutable-radix
-https://github.com/hashicorp/go-immutable-radix/blob/master/LICENSE
-
-hashicorp/golang-lru (Mozilla Public License 2.0) https://github.com/hashicorp/golang-lru
-https://github.com/hashicorp/golang-lru/blob/master/LICENSE
-
-gomodule/redigo (Apache 2.0) https://github.com/gomodule/redigo
-https://github.com/gomodule/redigo/blob/master/LICENSE
-
-OneOfOne/xxhash (Apache 2.0) https://github.com/OneOfOne/xxhash
-https://github.com/OneOfOne/xxhash/blob/master/LICENSE
-
-imdario/mergo (BSD-3) github.com/imdario/mergo
-https://github.com/imdario/mergo/blob/master/LICENSE
-
-magiconair/properties (BSD-2) https://github.com/magiconair/properties
-https://github.com/magiconair/properties/blob/master/LICENSE
-
-gopkg.in/eapache/queue.v1 (MIT) gopkg.in/eapache/queue.v1
-https://github.com/eapache/queue/blob/v1.1.0/LICENSE
-
-bertimus9/systemstat (MIT) https://bitbucket.org/bertimus9/systemstat
-https://bitbucket.org/bertimus9/systemstat/src/master/LICENSE
-
-davecgh/go-spew (ISC) https://github.com/davecgh/go-spew
-https://github.com/davecgh/go-spew/blob/master/LICENSE
-
-edgexfoundry/go-mod-core-contracts (Apache 2.0) https://github.com/edgexfoundry/go-mod-core-contracts
-https://github.com/edgexfoundry/go-mod-core-contracts/blob/master/LICENSE
-
-edgexfoundry/go-mod-registry (Apache 2.0) https://github.com/edgexfoundry/go-mod-registry
-https://github.com/edgexfoundry/go-mod-registry/blob/master/LICENSE
-
-edgexfoundry/go-mod-messaging (Apache 2.0) https://github.com/edgexfoundry/go-mod-messaging
-https://github.com/edgexfoundry/go-mod-messaging/blob/master/LICENSE
-
-gorilla/context (BSD-3) https://github.com/gorilla/context
-https://github.com/gorilla/context/blob/master/LICENSE
-
-kr/logfmt (Unspecified) https://github.com/kr/logfmt
-https://github.com/kr/logfmt/blob/master/Readme
-
-pmezard/go-difflib (Unspecified) https://github.com/pmezard/go-difflib
-https://github.com/pmezard/go-difflib/blob/master/LICENSE
-
-stretchr/objx (MIT) https://github.com/stretchr/objx
-https://github.com/stretchr/objx/blob/master/LICENSE
-
-stretchr/testify (MIT) https://github.com/stretchr/testify
-https://github.com/stretchr/testify/blob/master/LICENSE
-
-ugorji/go (MIT) https://github.com/ugorji/go
-https://github.com/ugorji/go/blob/master/LICENSE
-
-golang.org/x/net (Unspecified) https://github.com/golang/net
-https://github.com/golang/net/blob/master/LICENSE
-
-gopkg.in/yaml.v2 (Apache 2.0) https://github.com/go-yaml/yaml/
-https://github.com/go-yaml/yaml/blob/v2.2.2/LICENSE
